Moved mocks to Mocks.swift file
