+++ Sources/Bugger/UI/BuggerScreenshotCarousel.swift
@@ -208,36 +208,6 @@ extension BuggerScreenshotSource {
             loadFromManual: { [] }
         )
     }
-
-    static var previewMock: BuggerScreenshotSource {
-        BuggerScreenshotSource(
-            mode: .manual,
-            addTitle: "Add Sample",
-            loadFromPicker: { _ in [] },
-            loadFromManual: {
-                await MainActor.run {
-                    [
-                        makeSample(color: .systemBlue),
-                        makeSample(color: .systemGreen),
-                        makeSample(color: .systemOrange)
-                    ].compactMap { data in
-                        guard let data else { return nil }
-                        return BugReportAttachment(data: data, mimeType: "image/png")
-                    }
-                }
-            }
-        )
-    }
-
-    private static func makeSample(color: UIColor) -> Data? {
-        let size = CGSize(width: 300, height: 600)
-        let renderer = UIGraphicsImageRenderer(size: size)
-        let image = renderer.image { context in
-            color.setFill()
-            context.fill(CGRect(origin: .zero, size: size))
-        }
-        return image.pngData()
-    }
 }
 
 struct BuggerScreenshotThumbnail: View {
